change to partial capacity orders

+++ src/main/java/com/weekday/orderingsystem/ov1/dto/Order.java
@@ -19,5 +19,5 @@ import lombok.NoArgsConstructor;
 public class Order {
     int orderId;
     List<Character> meals;
-    int distance;
+    double distance;
 }
